Updates:
- updates "develooper_sample_posts_delete()" function to remove only sample posts.

+++ functions/develooper_sample_posts_delete.php
@@ -14,35 +14,30 @@ if ( ! defined('ABSPATH')) {
 
 require_once LOOPIS_DEV_DIR .'functions/sample.php';
 
+/**
+ * Delete sample posts from wp_posts
+ * 
+ * @return void
+ */
 function develooper_sample_posts_delete() {
 
     loopis_elog_function_start('develooper_sample_posts_delete');
     $sample_posts_info = get_sample_posts();
-    $posts = get_posts(['numberposts' => -1]);
-
-    error_log("Length: " . count($posts));
-
     
     global $wpdb;
     // Get posts
-    foreach ($posts as $post) {
-      // Delete all posts
-      wp_delete_post($post->ID, true);
+    foreach ($sample_posts_info as $sample_post) {
+        $post_name = $sample_post['post_name'];
+        $post = get_page_by_path($post_name, OBJECT, 'post');
+        if ($post) {
+            wp_delete_post($post->ID, true);
+            error_log("Deleted post: " . $post_name);
+        } else {
+            error_log("Post not found: " . $post_name);
+        }
     }
 
     // Resets post count
     $wpdb->query("ALTER TABLE {$wpdb->posts} AUTO_INCREMENT = 1");
     loopis_elog_function_end_success('develooper_sample_posts_delete');
 }
-
-/**
- * global $wpdb; Get all 
- * users except user 1(admin)
-   * $users = get_users(['exclude' => [1]]);
-    *foreach ($users as $user) {
-     *   // Delete each user
-      *  wp_delete_user($user->ID);
-    *}
-    *  //Resets user count
-    *$wpdb->query("ALTER TABLE {$wpdb->users} AUTO_INCREMENT = 1");
- */
